Apache Factory Toolkit: Wip if source file exist.

+++ wipe.py
@@ -1,3 +1,4 @@
+import os
 import sys
 
 source_file = ""
@@ -13,24 +14,25 @@ for arg in sys.argv:
     if index >= 3:
         replacements.append(arg)
 
-print("Wiping:")
-print("From: " + source_file)
-print("Into: " + destination_file)
+if os.path.isfile(source_file):
+    print("Wiping:")
+    print("From: " + source_file)
+    print("Into: " + destination_file)
 
-replace_what = []
-replace_with = []
-for x in range(0, replacements.__len__()):
-    if x % 2 == 0:
-        replace_what.append(replacements[x])
-    else:
-        replace_with.append(replacements[x])
+    replace_what = []
+    replace_with = []
+    for x in range(0, replacements.__len__()):
+        if x % 2 == 0:
+            replace_what.append(replacements[x])
+        else:
+            replace_with.append(replacements[x])
 
-with open(source_file, "rt") as fin:
-    with open(destination_file, "wt") as fout:
-        for line in fin:
-            replaced = line
-            for x in range(0, replacements.__len__() / 2):
-                replaced = replaced.replace(replace_what[x], replace_with[x])
-            fout.write(replaced)
+    with open(source_file, "rt") as fin:
+        with open(destination_file, "wt") as fout:
+            for line in fin:
+                replaced = line
+                for x in range(0, replacements.__len__() / 2):
+                    replaced = replaced.replace(replace_what[x], replace_with[x])
+                fout.write(replaced)
 
-print("-----")
+    print("-----")
